+++ examples/pre-training/ernie/src/utils/__init__.py
@@ -14,5 +14,4 @@
 
 
 from .logging import *  # noqa
-from .model_io import *  # noqa
 from .seed_utils import *  # noqa
