--- examples/pre-training/ernie/src/utils/model_io.py
@@ -1,30 +1,0 @@
-# Copyright (c) 2025 PaddlePaddle Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import paddle
-import numpy as np
-from .logging import logger
-
-
-def load_model(model, path, dtype=None):
-    sd = paddle.load(str(path), return_numpy=True)
-    for k, v in sd.items():
-        if k.endswith("position_ids"):
-            sd[k] = v.astype(np.int64)
-        if dtype and str(v.dtype) != dtype:
-            sd[k] = v.astype(dtype)
-    missing, unexpected = model.set_state_dict(sd)
-    logger.warning(
-        f"loading ckpt {path} done, unexpected={unexpected}, missing={missing}"
-    )
