+++ examples/pre-training/models/comm_utils.py
@@ -20,19 +20,10 @@ import numpy as np
 import paddle
 from paddle import distributed as dist
 from paddle.distributed import fleet
-from paddle.distributed.communication.batch_isend_irecv import (
-    _coalescing_manager as batch_isend_irecv_coalescing_manager,
-)
 from paddle.nn import functional as F
 from paddleformers.trainer.plugins.timer import get_timers
 
 logger = logging.getLogger(__name__)
-
-
-def md5(tensor):
-    numpy_array = tensor.numpy()
-    array_bytes = numpy_array.tobytes()
-    return hashlib.md5(array_bytes).hexdigest()
 
 
 def scatter(input, group=None, axis=0):
@@ -45,10 +36,13 @@ def scatter(input, group=None, axis=0):
     rank = group.rank
     seq_len = input.shape[axis]
     assert seq_len % parallelism == 0, (
-        f"Input sequence length {seq_len} can't be divided exactly" f" by sequence parallelism {parallelism}"
+        f"Input sequence length {seq_len} can't be divided exactly"
+        f" by sequence parallelism {parallelism}"
     )
     interval = seq_len // parallelism
-    input = paddle.slice(input, axes=[axis], starts=[interval * rank], ends=[interval * (rank + 1)])
+    input = paddle.slice(
+        input, axes=[axis], starts=[interval * rank], ends=[interval * (rank + 1)]
+    )
     input = paddle.assign(input)
     return input
 
@@ -65,7 +59,9 @@ def mp_slice(x, indices=None, group=None, axis=0):
     rank = group.rank
     assert len(indices) == parallelism, (len(indices), parallelism)
     indices = F.pad(paddle.to_tensor(indices).cumsum(0), [1, 0])
-    input = paddle.slice(x, axes=[axis], starts=[indices[rank]], ends=[indices[rank + 1]])
+    input = paddle.slice(
+        x, axes=[axis], starts=[indices[rank]], ends=[indices[rank + 1]]
+    )
     input = paddle.assign(input)
     return input
 
@@ -127,7 +123,9 @@ def all_gather(input, group=None, axis=0):
         output = paddle.empty(shape=output_shape, dtype=input.dtype)
         dist.stream.all_gather(output, input, group=group, use_calc_stream=True)
         return output
-    outputs = [paddle.empty(output_shape, dtype=input.dtype) for _ in range(parallelism)]
+    outputs = [
+        paddle.empty(output_shape, dtype=input.dtype) for _ in range(parallelism)
+    ]
     dist.stream.all_gather(outputs, input, group=group, use_calc_stream=True)
     output = paddle.concat(outputs, axis=axis)
     return output
@@ -146,7 +144,9 @@ def reduce_scatter(input, group=None):
     ), f"Input sequence length {input.shape[0]} can't be divided exactly by sequence parallelism {parallelism}"
     output_shape[0] = output_shape[0] // parallelism
     output = paddle.empty(shape=output_shape, dtype=input.dtype)
-    dist.stream.reduce_scatter(output, input, op=dist.ReduceOp.SUM, group=group, use_calc_stream=True)
+    dist.stream.reduce_scatter(
+        output, input, op=dist.ReduceOp.SUM, group=group, use_calc_stream=True
+    )
     return output
 
 
@@ -154,7 +154,9 @@ def subbatch(f, arg_idx, axis, bs, out_idx, use_recompute=False, same_arg_idx={}
     @functools.wraps(f)
     def wrapper(*args, **kwargs):
 
-        assert len(arg_idx) == len(axis), "Number of batching args and number of batching dims should match."
+        assert len(arg_idx) == len(
+            axis
+        ), "Number of batching args and number of batching dims should match."
 
         inps = [args[i] for i in arg_idx]
         axis_width = [inp.shape[d] for inp, d in zip(inps, axis)]
@@ -193,105 +195,6 @@ def subbatch(f, arg_idx, axis, bs, out_idx, use_recompute=False, same_arg_idx={}
         return paddle.concat(outs, out_idx)
 
     return wrapper
-
-
-def gather_varlen(input, dst, group, offload_pp_data_chunk_size=0, all_shape_and_dtype=None):
-    if dist.get_world_size(group) <= 1:
-        return input
-    if group is None:
-        group = dist.collective._get_global_group()
-
-    shape_and_dtype = (None, None) if input is None else (input.shape, input.dtype)
-    if all_shape_and_dtype is None:
-        all_shape_and_dtype = []
-        dist.all_gather_object(all_shape_and_dtype, shape_and_dtype, group=group)
-    assert any(s is not None for s, _ in all_shape_and_dtype), all_shape_and_dtype
-
-    any_shape = None
-    shape0_all = []
-    for s, d in all_shape_and_dtype:
-        if s is not None and any_shape is None:
-            any_shape = s
-        elif s is not None and any_shape is not None:
-            assert any_shape[1:] == s[1:], f"{any_shape[1:]} != {s[1:]}"
-        shape0_all.append(s if s is not None else 0)
-
-    output = []
-    if offload_pp_data_chunk_size > 0:
-        assert (group.nranks >= offload_pp_data_chunk_size) and (group.nranks % offload_pp_data_chunk_size == 0), (
-            f"group.nranks {group.nranks} must be greater than offload_pp_data_chunk_size {offload_pp_data_chunk_size} "
-            f"and group.nranks % offload_pp_data_chunk_size == 0"
-        )
-        if group.ranks[group.rank] == dst:
-            num_sub_group = group.nranks // offload_pp_data_chunk_size
-            for sub_group_idx in range(num_sub_group):
-                start = sub_group_idx * offload_pp_data_chunk_size
-                end = start + offload_pp_data_chunk_size
-                tasks = []
-                output_ptr = len(output)
-                with batch_isend_irecv_coalescing_manager(group, tasks):
-                    for src in range(start, end):
-                        if all_shape_and_dtype[src][0] is None or all_shape_and_dtype[src][0][0] == 0:
-                            pass
-                        elif src != group.rank:
-                            recv_tensor = paddle.empty(
-                                all_shape_and_dtype[src][0],
-                                dtype=all_shape_and_dtype[src][1],
-                            )
-                            output.append(recv_tensor)
-                            task = dist.irecv(recv_tensor, group.ranks[src], group=group)
-                            tasks.append(task)
-                        else:
-                            output.append(input)
-                    for task in tasks:
-                        task.wait()
-                for i in range(output_ptr, len(output)):
-                    output[i] = output[i].pin_memory()
-        else:
-            num_sub_group = group.nranks // offload_pp_data_chunk_size
-            for sub_group_idx in range(num_sub_group):
-                start = sub_group_idx * offload_pp_data_chunk_size
-                end = start + offload_pp_data_chunk_size
-                tasks = []
-                with batch_isend_irecv_coalescing_manager(group, tasks):
-                    for _ in range(1):
-                        if group.rank in list(range(start, end)) and input is not None and input.shape[0] != 0:
-                            task = dist.isend(input, dst, group=group)
-                            tasks.append(task)
-                for task in tasks:
-                    task.wait()
-    else:
-        if group.ranks[group.rank] == dst:
-            tasks = []
-            with batch_isend_irecv_coalescing_manager(group, tasks):
-                for src in range(group.nranks):
-                    if all_shape_and_dtype[src][0] is None:
-                        pass
-                    elif src != group.rank:
-                        recv_tensor = paddle.empty(
-                            all_shape_and_dtype[src][0],
-                            dtype=all_shape_and_dtype[src][1],
-                        )
-                        output.append(recv_tensor)
-                        task = dist.irecv(recv_tensor, group.ranks[src], group=group)
-                        tasks.append(task)
-                    else:
-                        output.append(input)
-            for task in tasks:
-                task.wait()
-        else:
-            tasks = []
-            with batch_isend_irecv_coalescing_manager(group, tasks):
-                for _ in range(1):
-                    if input is not None:
-                        task = dist.isend(input, dst, group=group)
-                        tasks.append(task)
-            for task in tasks:
-                task.wait()
-
-        if len(output) != 0:
-            output = paddle.concat(output, 0)
-    return output
 
 
 @contextmanager
