+++ examples/pre-training/models/moe/__init__.py
@@ -13,11 +13,9 @@
 # limitations under the License.
 
 from .moe_layer import MOELayer
-from .round_robin_gate import RoundRobinGate
 from .top2_gate import Top2Gate
 
 __all__ = [
     "MOELayer",
-    "RoundRobinGate",
     "Top2Gate",
 ]
